Move linearCombination and customColvar to colvarcomp_combination.cpp
This commit moves the linearCombination and customColvar to
colvarcomp_combination.cpp. In addition, to avoid the redundant
comparison function compareColvarComponent in colvarcomp_gpath.cpp, the
compare function in colvar.cpp is moved to be a static member function
of the colvar class. Then linearCombination and customColvar reuse
colvar::compare_cvc for sorting.
